+++ raytracer_weekend_lib/src/hittable/rectangular.rs
@@ -15,16 +15,16 @@ use crate::{
 
 #[derive(Debug, Constructor)]
 pub struct XYRectangle {
-    x0: f64,
-    x1: f64,
-    y0: f64,
-    y1: f64,
-    k: f64,
+    x0: f32,
+    x1: f32,
+    y0: f32,
+    y1: f32,
+    k: f32,
     material: Box<dyn Material>,
 }
 
 impl Hittable for XYRectangle {
-    fn hit(&self, r: &Ray, t_min: f64, t_max: f64, _rng: &mut ActiveRng) -> Option<HitRecord> {
+    fn hit(&self, r: &Ray, t_min: f32, t_max: f32, _rng: &mut ActiveRng) -> Option<HitRecord> {
         let x0 = self.x0;
         let y0 = self.y0;
         let x1 = self.x1;
@@ -56,7 +56,7 @@ impl Hittable for XYRectangle {
         ));
     }
 
-    fn bounding_box(&self, _time0: f64, _time1: f64) -> Option<Aabb> {
+    fn bounding_box(&self, _time0: f32, _time1: f32) -> Option<Aabb> {
         Some(Aabb::new(
             Point3::new(self.x0, self.y0, self.k - 0.0001),
             Point3::new(self.x1, self.y1, self.k + 0.0001),
@@ -66,16 +66,16 @@ impl Hittable for XYRectangle {
 
 #[derive(Debug, Constructor)]
 pub struct XZRectangle {
-    x0: f64,
-    x1: f64,
-    z0: f64,
-    z1: f64,
-    k: f64,
+    x0: f32,
+    x1: f32,
+    z0: f32,
+    z1: f32,
+    k: f32,
     material: Box<dyn Material>,
 }
 
 impl Hittable for XZRectangle {
-    fn hit(&self, r: &Ray, t_min: f64, t_max: f64, _rng: &mut ActiveRng) -> Option<HitRecord> {
+    fn hit(&self, r: &Ray, t_min: f32, t_max: f32, _rng: &mut ActiveRng) -> Option<HitRecord> {
         let x0 = self.x0;
         let z0 = self.z0;
         let x1 = self.x1;
@@ -107,7 +107,7 @@ impl Hittable for XZRectangle {
         ));
     }
 
-    fn bounding_box(&self, _time0: f64, _time1: f64) -> Option<Aabb> {
+    fn bounding_box(&self, _time0: f32, _time1: f32) -> Option<Aabb> {
         Some(Aabb::new(
             Point3::new(self.x0, self.k - 0.0001, self.z0),
             Point3::new(self.x1, self.k + 0.0001, self.z1),
@@ -117,16 +117,16 @@ impl Hittable for XZRectangle {
 
 #[derive(Debug, Constructor)]
 pub struct YZRectangle {
-    y0: f64,
-    y1: f64,
-    z0: f64,
-    z1: f64,
-    k: f64,
+    y0: f32,
+    y1: f32,
+    z0: f32,
+    z1: f32,
+    k: f32,
     material: Box<dyn Material>,
 }
 
 impl Hittable for YZRectangle {
-    fn hit(&self, r: &Ray, t_min: f64, t_max: f64, _rng: &mut ActiveRng) -> Option<HitRecord> {
+    fn hit(&self, r: &Ray, t_min: f32, t_max: f32, _rng: &mut ActiveRng) -> Option<HitRecord> {
         let y0 = self.y0;
         let z0 = self.z0;
         let y1 = self.y1;
@@ -158,7 +158,7 @@ impl Hittable for YZRectangle {
         ));
     }
 
-    fn bounding_box(&self, _time0: f64, _time1: f64) -> Option<Aabb> {
+    fn bounding_box(&self, _time0: f32, _time1: f32) -> Option<Aabb> {
         Some(Aabb::new(
             Point3::new(self.k - 0.0001, self.y0, self.z0),
             Point3::new(self.k + 0.0001, self.y1, self.z1),
@@ -235,11 +235,11 @@ impl Cuboid {
 }
 
 impl Hittable for Cuboid {
-    fn hit(&self, r: &Ray, t_min: f64, t_max: f64, rng: &mut ActiveRng) -> Option<HitRecord> {
+    fn hit(&self, r: &Ray, t_min: f32, t_max: f32, rng: &mut ActiveRng) -> Option<HitRecord> {
         self.sides.hit(r, t_min, t_max, rng)
     }
 
-    fn bounding_box(&self, _time0: f64, _time1: f64) -> Option<Aabb> {
+    fn bounding_box(&self, _time0: f32, _time1: f32) -> Option<Aabb> {
         Some(Aabb::new(self.box_min, self.box_max))
     }
 }
