+++ raytracer_weekend_lib/src/hittable/volumes.rs
@@ -1,5 +1,7 @@
 use core::fmt::Debug;
 
+#[cfg(feature = "no_std")]
+use micromath::F32Ext;
 use rand::Rng;
 
 use crate::{
@@ -16,11 +18,11 @@ use crate::{
 pub struct ConstantMedium<H: Hittable, T: Texture> {
     boundary: H,
     phase_function: Isotropic<T>,
-    neg_inv_density: f64,
+    neg_inv_density: f32,
 }
 
 impl<H: Hittable, T: Texture> ConstantMedium<H, T> {
-    pub fn new(boundary: H, density: f64, texture: T) -> Self {
+    pub fn new(boundary: H, density: f32, texture: T) -> Self {
         let neg_inv_density = -1.0 / density;
         let phase_function = Isotropic::new(texture);
 
@@ -33,11 +35,11 @@ impl<H: Hittable, T: Texture> ConstantMedium<H, T> {
 }
 
 impl<H: Hittable, T: Texture> Hittable for ConstantMedium<H, T> {
-    fn hit(&self, r: &Ray, t_min: f64, t_max: f64, rng: &mut ActiveRng) -> Option<HitRecord> {
+    fn hit(&self, r: &Ray, t_min: f32, t_max: f32, rng: &mut ActiveRng) -> Option<HitRecord> {
         let rec1 = self
             .boundary
-            .hit(r, f64::NEG_INFINITY, f64::INFINITY, rng)?;
-        let rec2 = self.boundary.hit(r, rec1.t + 0.0001, f64::INFINITY, rng)?;
+            .hit(r, f32::NEG_INFINITY, f32::INFINITY, rng)?;
+        let rec2 = self.boundary.hit(r, rec1.t + 0.0001, f32::INFINITY, rng)?;
 
         let mut rec1_t = rec1.t;
         let mut rec2_t = rec2.t;
@@ -53,7 +55,7 @@ impl<H: Hittable, T: Texture> Hittable for ConstantMedium<H, T> {
 
         let ray_length = r.direction().length();
         let distance_inside_boundary = (rec2_t - rec1_t) * ray_length;
-        let hit_distance = self.neg_inv_density * rng.gen::<f64>().log10();
+        let hit_distance = self.neg_inv_density * rng.gen::<f32>().log10();
 
         if hit_distance > distance_inside_boundary {
             return None;
@@ -75,7 +77,7 @@ impl<H: Hittable, T: Texture> Hittable for ConstantMedium<H, T> {
         ))
     }
 
-    fn bounding_box(&self, time0: f64, time1: f64) -> Option<Aabb> {
+    fn bounding_box(&self, time0: f32, time1: f32) -> Option<Aabb> {
         self.boundary.bounding_box(time0, time1)
     }
 }
