+++ raytracer_weekend_lib/src/ray.rs
@@ -6,7 +6,7 @@ use super::vec3::{Point3, Vec3};
 pub struct Ray {
     origin: Point3,
     direction: Vec3,
-    time: f64,
+    time: f32,
 }
 
 impl Ray {
@@ -18,11 +18,11 @@ impl Ray {
         self.direction
     }
 
-    pub fn time(&self) -> f64 {
+    pub fn time(&self) -> f32 {
         self.time
     }
 
-    pub fn at(&self, t: f64) -> Point3 {
+    pub fn at(&self, t: f32) -> Point3 {
         self.origin + t * self.direction
     }
 }
